Tweaks, and fixes to uniform compatibilities
