+++ addons/cdf/CfgEditorSubcategories.hpp
@@ -1,0 +1,5 @@
+class CfgEditorSubcategories {
+    class GVAR(men_woodland) {
+        displayName = "Men (Woodland)";
+    };
+};
