+++ addons/cdf/vehicles/CfgRG31.hpp
@@ -15,7 +15,7 @@ class CUP_B_RG31_M2_OD_USA;
 class GVAR(RG31_HMG_Woodland) : CUP_B_RG31_M2_OD_USA {
     side = 0;
     faction = QUOTE(ADDON);
-    displayName = "RG31 HMG (Woodland)";
+    displayName = "RG31 HMG";
     crew = QGVAR(Rifleman_Woodland);
     typicalCargo[] = {
         QGVAR(Rifleman_Woodland)
@@ -40,7 +40,7 @@ class CUP_B_RG31_Mk19_OD_USA;
 class GVAR(RG31_GMG_Woodland) : CUP_B_RG31_Mk19_OD_USA {
     side = 0;
     faction = QUOTE(ADDON);
-    displayName = "RG31 GMG (Woodland)";
+    displayName = "RG31 GMG";
     crew = QGVAR(Rifleman_Woodland);
     typicalCargo[] = {
         QGVAR(Rifleman_Woodland)
@@ -65,7 +65,7 @@ class CUP_B_RG31E_M2_OD_USA;
 class GVAR(RG31_Tpt_Woodland) : CUP_B_RG31E_M2_OD_USA {
     side = 0;
     faction = QUOTE(ADDON);
-    displayName = "RG31 Troop Carrier (Woodland)";
+    displayName = "RG31 Troop Carrier";
     crew = QGVAR(Rifleman_Woodland);
     typicalCargo[] = {
         QGVAR(Rifleman_Woodland)
